changed dtype

+++ models/GCN_model.py
@@ -10,7 +10,7 @@ from layers import *
 use_cuda = torch.cuda.is_available()
 FloatTensor = torch.cuda.FloatTensor if use_cuda else torch.FloatTensor
 device = torch.device('cuda') if use_cuda else torch.device('cpu')
-dtype = torch.float32
+dtype = torch.float64
 torch.set_default_tensor_type(FloatTensor)
 
 def Normalize_Adj(A):
